refactor: HttpRequestProcessor 추상화
